+++ practical/ProcessMining/group2/inductive_miner/data/images/loop_arrow.png
