+++ practical/ProcessMining/group2/inductive_miner/data/images/seq_arrow.png
