+++ practical/ProcessMining/group2/inductive_miner/data/images/x.png
