+++ practical/ProcessMining/group2/inductive_miner/output/pm4py_process_trees.txt
