+++ practical/ProcessMining/group2/inductive_miner/src/process_tree.py
@@ -1,74 +1,10 @@
-import os
 import copy
 import itertools
-from practical.ProcessMining.group2.inductive_miner.src.graph_utils import *
 
-script_dir = os.path.dirname(os.path.abspath(__file__))
+from practical.ProcessMining.group2.inductive_miner.src.graph import *
+from practical.ProcessMining.group2.inductive_miner.src.event_log import *
+from practical.ProcessMining.group2.inductive_miner.src.directly_follows_graph import *
 
-class EventLog:
-    def __init__(self, traces: Dict[str, int]):
-        """
-        Initialize EventLog object.
-
-        :param traces: Dictionary where keys are traces and values are counts.
-        """
-        self.traces = traces
-
-    @classmethod
-    def from_file(cls, file_path: str = None) -> 'EventLog':
-        """
-        Create an EventLog object from a file.
-
-        :param file_path: Path to the file containing traces.
-        :return: EventLog object.
-        """
-        with open(os.path.join(script_dir, file_path), 'r') as file:
-            traces = {}
-            for line in file:
-                trace = line.strip()
-                if trace in traces:
-                    traces[trace] += 1
-                else:
-                    traces[trace] = 1
-            return EventLog(traces)
-
-class DirectlyFollowsGraph(Graph):
-    def __init__(self, event_log: EventLog) -> None:
-        """
-        Initialize DirectlyFollowsGraph object.
-
-        :param event_log: EventLog object.
-        """
-        self.event_log = event_log
-        self.graph = defaultdict(list)  # Adjacency list, format: "node_id" : [children]
-        self.start_nodes = set()
-        self.end_nodes = set()
-
-    def construct_dfg(self) -> None:
-        """Construct Directly Follows Graph."""
-        for trace in self.event_log.traces.keys():
-            if trace:  # Check if trace is not empty
-                self.start_nodes.add(trace[0])  # First activity in trace is a start
-                self.end_nodes.add(trace[-1])  # Last activity in trace is an end
-
-                for i in range(len(trace) - 1):
-                    current_activity = trace[i]
-                    next_activity = trace[i + 1]
-
-                    if next_activity not in self.graph[current_activity]:
-                        self.graph[current_activity].append(next_activity)
-                    if (next_activity not in self.graph):  # Add next activity to graph if not already present, i.e. if it is an end node
-                        self.graph[next_activity] = []
-                
-                for node in self.start_nodes:
-                    if node not in self.graph: # Isolated nodes
-                        self.graph[node] = []
-            else:
-                self.graph[''] = [] # Add tau node for empty traces
-
-    def __str__(self) -> str:
-        """Represent Directly Follows Graph as string."""
-        return f"Directly Follows Graph: (\n\tGraph: {dict(self.graph)}\n\tStart nodes: {self.start_nodes}\n\tEnd nodes: {self.end_nodes}\n)"
 
 class ProcessTree:
     def __init__(self, event_log: EventLog) -> None:
@@ -508,20 +444,4 @@ class ProcessTree:
                 return f"{operator_map[operator]}({', '.join(print_tree(child) for child in children)})"
             
         return print_tree(tree)
-            
-class InductiveMiner():
-    def mine_process_model(self, event_log: EventLog) -> ProcessTree:
-        """
-        Mine process model using Inductive Miner.
-
-        :param event_log: EventLog object.
-        :return: ProcessTree object.
-        """
-        # Construct Directly-Follows Graph (DFG)
-        dfg = DirectlyFollowsGraph(event_log)
-        dfg.construct_dfg()
-
-        # Construct Process Tree from DFG
-        process_tree = ProcessTree(event_log)
-
-        return process_tree
+
